+++ .storybook/preview.ts
@@ -15,9 +15,9 @@ import "@fontsource/material-icons";
 
 const customViewports = {
   small: {
-    name: "300",
+    name: "320",
     styles: {
-      width: "300px",
+      width: "320px",
       height: "700px",
     },
   },
@@ -38,7 +38,7 @@ const customViewports = {
   desktop: {
     name: "960",
     styles: {
-      width: "600px",
+      width: "960px",
       height: "700px",
     },
   },
@@ -85,7 +85,14 @@ const preview: Preview = {
     // },
     viewport: {
       viewports: { ...customViewports, ...INITIAL_VIEWPORTS },
-      defaultViewport: "960",
+      defaultViewport: "desktop",
+    },
+    playroom: {
+      url:
+        process.env.NODE_ENV === "development"
+          ? "http://localhost:9000"
+          : "http://urt-web-app1:3047/playroom",
+      code: "code here ...",
     },
   },
 };