+++ serve.js
@@ -3,16 +3,9 @@ const path = require("path");
 
 // Constants
 const PORT = 8080;
-const HOST = "0.0.0.0";
 
 // App
 const app = express();
-app.use(express.static("storybook-static"));
+app.use(express.static("public"));
 
-app.get("/", (req, res) => {
-  res.sendFile(path.join("storybook-static", "index.html"));
-});
-
-app.listen(PORT, HOST);
-
-console.log(`Running on http://${HOST}:${PORT}`);
+app.listen(PORT);