+++ src/components/Notification/Notification.stories.tsx
@@ -7,6 +7,15 @@ const meta: Meta<typeof Notification> = {
   title: "Components/Notification",
   component: Notification,
   tags: ["autodocs"],
+  parameters: {
+    playroom: {
+      code: `
+        <MainContainer>
+          <Notification variant="reject"/>
+        </MainContainer>
+      `,
+    },
+  },
 };
 
 export default meta;
@@ -17,16 +26,43 @@ export const Reject: Story = {
   args: {
     variant: "reject",
   },
+  parameters: {
+    playroom: {
+      code: `
+        <MainContainer>
+          <Notification variant="reject"/>
+        </MainContainer>
+      `,
+    },
+  },
 };
 
 export const TechError: Story = {
   args: {
     variant: "techError",
   },
+  parameters: {
+    playroom: {
+      code: `
+        <MainContainer>
+          <Notification variant='techError'/>
+        </MainContainer>
+      `,
+    },
+  },
 };
 
 export const WeRecognizedYou: Story = {
   args: {
     variant: "weRecognizedYou",
+  },
+  parameters: {
+    playroom: {
+      code: `
+      <MainContainer>
+        <Notification variant='weRecognizedYou'/>
+      </MainContainer>
+    `,
+    },
   },
 };